requirements.txt
selenium 
psycopg2 
flask
python-dotenv